Refactor userMenu Context
