+++ src/components/AppWrapper/AppWrapper.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { makeStyles, useTheme, Theme, createStyles } from '@material-ui/core/styles';
 import SwipeableDrawer from '@material-ui/core/SwipeableDrawer';
-import MenuContext from 'contexts/MenuContext';
+import { useMenu } from 'contexts/MenuContext';
 import UserMenu from 'components/UserMenu';
 import { convertDirectionToAnchor } from 'utils';
 import SnackBar from 'components/SnackBar';
@@ -32,8 +32,7 @@ const AppWrapper: React.FC = props => {
   const { children } = props;
   const classes = useStyles();
   const theme = useTheme();
-  const { isOpen, toggleMenu } = React.useContext(MenuContext);
-
+  const { isOpen, toggleMenu } = useMenu();
   return (
     <div className={classes.root} data-testid='app-wrapper'>
       <nav>
