--- src/stores/MenuProvider.tsx
@@ -1,24 +1,0 @@
-import React, { useState } from 'react';
-import MenuContext, { defaultState } from 'contexts/MenuContext';
-
-interface MenuProviderProps {
-  children: React.ReactNode;
-}
-
-const MenuProvider = (props: MenuProviderProps): JSX.Element => {
-  const [state, toggleMenuState] = useState(defaultState);
-  const { isOpen } = state;
-  const toggleMenu = () => toggleMenuState({ ...state, isOpen: !isOpen });
-
-  return (
-    <MenuContext.Provider
-      value={{
-        isOpen,
-        toggleMenu,
-      }}>
-      {props.children}
-    </MenuContext.Provider>
-  );
-};
-
-export default MenuProvider;
